requirements.txt
Flask-SQLAlchemy
mutagen
pynput
requests